agent-work/requirements.txt
pydantic
boto3
strands-agents>=1.0.0
strands-agents-tools>=0.2.0